services/backend/requirements.txt
fastapi>=0.128.6
uvicorn>=0.40.0
pika
pandas
numpy
PyYAML
httpx>=0.27
pydantic>=2.0
